lore items only work while favorited

+++ CalPlayer/CalamityPlayerMiscEffects.cs
@@ -3844,7 +3844,7 @@ namespace CalamityMod.CalPlayer
 		#region Rogue Mirrors
 		private static void RogueMirrors(Player player, CalamityPlayer modPlayer)
 		{
-			Rectangle rectangle = new Rectangle((int)((double)player.position.X + (double)player.velocity.X * 0.5 - 4.0), (int)((double)player.position.Y + (double)player.velocity.Y * 0.5 - 4.0), player.width + 8, player.height + 8);
+			Rectangle rectangle = new Rectangle((int)(player.position.X + player.velocity.X * 0.5f - 4f), (int)(player.position.Y + player.velocity.Y * 0.5f - 4f), player.width + 8, player.height + 8);
 			for (int i = 0; i < Main.maxNPCs; i++)
 			{
 				NPC npc = Main.npc[i];
